# internal/dal/model/users.gen.go
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.

package model

const TableNameUser = "Users"

// User mapped from table <Users>
type User struct {
	ID   int64  `gorm:"column:Id;primaryKey;autoIncrement:true" json:"Id"`
	Name string `gorm:"column:Name" json:"Name"`
	Age  int32  `gorm:"column:age" json:"age"`
}

// TableName User's table name
func (*User) TableName() string {
	return TableNameUser
}